requirements.txt
requests
Pillow
weasyprint
markdown
